Herstructureer de MCP Document Processor door tools te centraliseren in een nieuw bestand. Pas de configuratie en serverinstellingen aan om de nieuwe toolstructuur te ondersteunen. Verander de naam van de processor van 'MCP Invoice Processor' naar 'MCP Document Processor' in de relevante bestanden. Verbeter de HTTP server door gebruik te maken van de gedeelde tools voor documentverwerking en health checks.
